feat: condition rendering buttons in header
