+++ src/components/Boards/Boards.tsx
@@ -3,11 +3,7 @@ import React, { useState } from 'react';
 import styles from './Boards.module.css';
 
 function Boards(): JSX.Element {
-  
-
-  return (
-    
-  );
+  return <div className="">hello</div>;
 }
 
 export default Boards;
