+++ src/components/Header/Header.tsx
@@ -1,10 +1,28 @@
 import React from 'react';
-import { Link } from 'react-router-dom';
+import { Link, useNavigate } from 'react-router-dom';
 import styles from './Header.module.css';
 import logo from '../../assets/img/Logo.png';
 import langchange from '../../assets/img/Langchange.png';
+import { useAuth } from '../hook/useAuth';
 
-function Header() {
+function Header({ logged = false }) {
+  const [state, dispatch] = useAuth();
+  const navigate = useNavigate();
+
+  function handleLogOut() {
+    localStorage.removeItem('token');
+    dispatch({
+      type: 'user',
+      data: {
+        username: null,
+        login: null,
+        token: null,
+        id: null,
+      },
+    });
+    navigate('/');
+  }
+
   return (
     <>
       <header>
@@ -15,12 +33,24 @@ function Header() {
           <div className={styles.nav__wrapper}>
             <img className={styles.lang} src={langchange} alt="choose lang" />
             <nav>
-              <Link to="/login">
-                <button className={styles.sign__in}>Sign In</button>
-              </Link>
-              <Link to="/register">
-                <button className={styles.sign__up}>Sign Up</button>
-              </Link>
+              {logged ? (
+                <Link to="/profile">
+                  <button className={styles.sign__in}>Edit Profile</button>
+                </Link>
+              ) : (
+                <Link to="/login">
+                  <button className={styles.sign__in}>Sign In</button>
+                </Link>
+              )}
+              {logged ? (
+                <button className={styles.sign__up} onClick={handleLogOut}>
+                  Log out
+                </button>
+              ) : (
+                <Link to="/register">
+                  <button className={styles.sign__up}>Sign Up</button>
+                </Link>
+              )}
             </nav>
           </div>
         </div>
