+++ src/index.css
@@ -23,7 +23,6 @@ a{
 a:visited{
   color: #000;
 }
-a:hover{color: #7795CA;}
 button{
   cursor: pointer;
   border: none;
